* Issue 25843 DocFactory - New Feature: Possibility to use a Serializable as datasource for the mergefields. (Supports embedded Serializable objects in Serializable). Example of use: DocumentTemplate documentTemplate = DocumentTemplate.withTemplate(template).putDataAsSourceForSimpleMailMerge(makePerson()); The mergefields will be named with the Serializable Class as prefix like: person.name person.firstname person.address.city.cityname
* Issue 25844 DocFactory - New Feature: Possibility to use a Serializable as datasource for the mergefields with region and nested merge fields with regions.
A java Bean may contain collections of other beans which in turn may contain other collections, and so on ... We have then nested collections inside other parent collections.
These collections may be used for injecting data in nested mail merge regions of your template.
IMPORTANT
For the moment cyclic relationships are not supported. For example an Order object has a list of OrderDetails objects, each OrderDetails object has a reference to the parent Order and so on... . 
For the moment the way the DocFactory introspects the data may leads to an OutOfMemory Error. This is a well known problem and this may be supported in the future.

Reviewed by ec

Former-commit-id: 2f07f0394386f790454ad991bd96b1c1e2118ae0